+++ CHANGELOG.md
@@ -132,6 +132,7 @@
 - Enh #133: Add `HasPatternContentUnits` trait and corresponding tests for managing SVG `patternContentUnits` attribute (@terabytesoftw)
 - Enh #134: Add `HasPatternTransform` trait and corresponding tests for managing SVG `patternTransform` attribute (@terabytesoftw)
 - Enh #135: Add `HasPatternUnits` trait and corresponding tests for managing SVG `patternUnits` attribute (@terabytesoftw)
+- Enh #136: Add `Pattern` class and corresponding tests for SVG `<pattern>` element functionality (@terabytesoftw)
 
 ## 0.2.0 March 31, 2024